Safari-609.1.11.1/LayoutTests/fast/events/webkit-media-key-events-constructor-expected.txt

Test the WebKitMediaKeyMessageEvent / WebKitMediaKeyNeededEvent constructors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


event = new WebKitMediaKeyMessageEvent("foo")
PASS event.type is "foo"
PASS event.bubbles is false
PASS event.cancelable is false
PASS event.message is null
PASS event.destinationURL is ""

event = new WebKitMediaKeyMessageEvent("foo", { bubbles: true, cancelable: true, message: testMessage, destinationURL: "testURL" })
PASS event.type is "foo"
PASS event.bubbles is true
PASS event.cancelable is true
PASS event.message is testMessage
PASS event.destinationURL is "testURL"

event = new WebKitMediaKeyNeededEvent("foo")
PASS event.type is "foo"
PASS event.bubbles is false
PASS event.cancelable is false
PASS event.initData is null

event = new WebKitMediaKeyNeededEvent("foo", { bubbles: true, cancelable: true, initData: testMessage })
PASS event.type is "foo"
PASS event.bubbles is true
PASS event.cancelable is true
PASS event.initData is testMessage

PASS successfullyParsed is true

TEST COMPLETE
